order api issue fixed push notification get profile api modified

+++ app/src/main/java/com/foodie/app/activities/CurrentOrderDetailActivity.java
@@ -122,7 +122,7 @@ public class CurrentOrderDetailActivity extends AppCompatActivity implements OnM
 
     Fragment orderFullViewFragment;
     FragmentManager fragmentManager;
-    int priceAmount = 0;
+    Double priceAmount = 0.0;
     int itemQuantity = 0;
     String currency = "";
     @BindView(R.id.order_flow_rv)
